Add files via upload
removed duplicate method from Attribute & Method classes
