+++ module_builder/attribute.py
@@ -1,10 +1,12 @@
+from .type_finder import TypeFinder
+
 class Attribute:
 
     """define the attributes of a class"""
 
     def __init__(self, new_name, new_type):
         self.name = new_name.replace(" ", "")
-        self.type = self.find_type(new_type)
+        self.type = TypeFinder.find_type(new_type)
 
     def __str__(self):
         if "str" in self.type:
@@ -17,16 +19,3 @@ class Attribute:
             return f"        self.{self.name}: {self.type} = ()  # ToDo\n"
         else:
             return f"        self.{self.name} = None  # ToDo\n"
-
-    @staticmethod
-    def find_type(new_type):
-        if "string" in new_type:
-            return "str"
-        elif "number" in new_type:
-            return "int"
-        elif "list" in new_type:
-            return "list"
-        elif "tuple" in new_type:
-            return "tuple"
-        else:
-            return ""
